requirements.txt
dask==2024.6.2
datasets==2.19.2
huggingface_hub==0.21.4
memory_profiler==0.61.0
numba==0.60.0
numpy>=1.16.6
pandas==1.5.3
prettytable
psutil==5.9.8
pyaedt>=0.6.46
pyarrow==15.0.1
pyEPR==1.1.5
pyEPR_quantum>=0.8.5.7
python-dotenv
scipy>=1.10.0
scqubits>=3.3.0
seaborn
setuptools
tabulate
tqdm
cython>=0.29.20
qutip>=4.3.1
addict
datashader>=0.16.0
joblib>=1.3.2
Requests==2.32.3
PyGithub==2.4.0
GitPython==3.1.43